# fp12: improved cyclotomicSquare